backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
package-mode = false
description = "Backend для DayTaskPlan"
# authors = [{name = "EvGen Gal",email = "ZBst.1@yandex.ru"}]
authors = ["EvGen Gal <ZBst.1@yandex.ru>"]
packages = [
    { include = "app" } 
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
uvicorn = "^0.34.0"
python-dotenv = "^1.1.0"
pydantic = {extras = ["dotenv"], version = "^2.11.1"}
passlib = "^1.7.4"
bcrypt = "^4.3.0"
celery = "^5.4.0"
redis = "^5.2.1"
websockets = "^15.0.1"
pydantic-settings = "^2.8.1"
psycopg2-binary = "^2.9.10"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[tool.poetry.scripts]
start-dev = "app.scripts.cmd.run_dev:main"
migrate-gen = "app.scripts.cmd.migrate_gen:main"
migrate-up = "app.scripts.cmd.migrate_up:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
